+++ samples/mcp/proxy/Dockerfile
@@ -1,0 +1,12 @@
+FROM node:18-alpine
+
+WORKDIR /app
+COPY package*.json tsconfig.json ./
+RUN npm install
+
+COPY . .
+
+RUN npm run build 
+
+EXPOSE 4000
+CMD ["npm", "run", "start"]
